Readme, lights, JSDocs
Added a readme to log all the changes and just my thoughts lol. I added light classes, material class, a lot of JDocs beacuse it makes me feel cool. Added the github page to the site and changed the title and a bit of styling for the link and the footer itself. I changed the main script to accomodate for the Scene class that will house all the properties in a scene. Also added materials to surfaces but I have not removed their color components just because I don't want to break that just yet. Still have to implement BRDF or whatever its called.